+++ src/main/java/com/blackbuild/groovy/configdsl/transform/DSL.java
@@ -1,0 +1,12 @@
+package com.blackbuild.groovy.configdsl.transform;
+
+import org.codehaus.groovy.transform.GroovyASTTransformationClass;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.SOURCE)
+@Inherited // This is currently not used, see https://issues.apache.org/jira/browse/GROOVY-6765
+@GroovyASTTransformationClass(classes={DSLASTTransformation.class})
+public @interface DSL {
+}