+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/model/DefaultValuesSpec.groovy
@@ -7,7 +7,7 @@ class DefaultValuesSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 String name
             }
@@ -37,7 +37,7 @@ class DefaultValuesSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 String name
                 String value
@@ -68,7 +68,7 @@ class DefaultValuesSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 String name
                 String value = "hallo"
@@ -91,7 +91,7 @@ class DefaultValuesSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 String name
                 String value
@@ -119,9 +119,9 @@ class DefaultValuesSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig(key = "name")
+            @DSL
             class Foo {
-                String name
+                @Key String name
                 String value
                 String value2
             }
@@ -149,7 +149,7 @@ class DefaultValuesSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<String> names
             }