+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/model/TransformSpec.groovy
@@ -12,7 +12,7 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
             }
         ''')
@@ -26,7 +26,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 boolean isCalled
                 Foo apply(Closure c) {
@@ -48,7 +48,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Foo apply(Closure c) { this }
                 Foo  _apply(Closure c) { this }
@@ -64,7 +64,7 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
             }
         ''')
@@ -81,7 +81,7 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
 
                 boolean called
@@ -106,9 +106,9 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig(key = "name")
+            @DSL
             class Foo {
-                String name
+                @Key String name
             }
         ''')
 
@@ -127,9 +127,9 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig(key = "name")
+            @DSL
             class Foo {
-                String name
+                @Key String name
                 boolean called
 
                 def static create(String key, Closure c) {
@@ -152,7 +152,7 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 def static create(Closure c) {
                     _create(c)
@@ -170,7 +170,7 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig(key="name")
+            @DSL
             class Foo {
                 String name
                 def static create(String key, Closure c) {
@@ -191,9 +191,9 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig(key = "name")
+            @DSL
             class Foo {
-                String name
+                @Key String name
             }
         ''')
 
@@ -205,28 +205,14 @@ class TransformSpec extends AbstractDSLSpec {
         instance.name == "Klaus"
     }
 
-    def "key field must exist"() {
-        when:
-        createClass('''
-            package pk
-
-            @DSLConfig(key = "name")
-            class Foo {
-            }
-        ''')
-
-        then:
-        thrown(MultipleCompilationErrorsException)
-    }
-
     def "key field must be of type String"() {
         when:
         createClass('''
             package pk
 
-            @DSLConfig(key = "name")
+            @DSL
             class Foo {
-                int name
+                @Key int name
             }
         ''')
 
@@ -237,7 +223,7 @@ class TransformSpec extends AbstractDSLSpec {
     def "simple member method"() {
         given:
         createInstance('''
-            @DSLConfig
+            @DSL
             class Foo {
                 String value
             }
@@ -255,12 +241,12 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Bar inner
             }
 
-            @DSLConfig
+            @DSL
             class Bar {
                 String name
             }
@@ -281,9 +267,9 @@ class TransformSpec extends AbstractDSLSpec {
     def "simple member method with renaming annotation"() {
         given:
         createInstance('''
-            @DSLConfig
+            @DSL
             class Foo {
-                @DSLField("firstname") String name
+                @Field("firstname") String name
                 String lastname
             }
         ''')
@@ -297,7 +283,7 @@ class TransformSpec extends AbstractDSLSpec {
     def "test existing method"() {
         given:
         createInstance('''
-            @DSLConfig
+            @DSL
             class Foo {
                 String name, lastname
                 def name(String value) {return "run"}
@@ -311,9 +297,9 @@ class TransformSpec extends AbstractDSLSpec {
     def "test existing method with renaming"() {
         given:
         createInstance('''
-            @DSLConfig
+            @DSL
             class Foo {
-                @DSLField("firstname") String name
+                @Field("firstname") String name
                 String lastname
                 def firstname(String value) {return "run"}
             }
@@ -328,12 +314,12 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Bar inner
             }
 
-            @DSLConfig
+            @DSL
             class Bar {
                 String name
             }
@@ -356,14 +342,14 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Bar inner
             }
 
-            @DSLConfig(key = "name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 int value
             }
         ''')
@@ -386,12 +372,12 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<Bar> bars
             }
 
-            @DSLConfig
+            @DSL
             class Bar {
                 String name
             }
@@ -408,17 +394,18 @@ class TransformSpec extends AbstractDSLSpec {
         instance.bars[1].name == "Klaus"
     }
 
+    @SuppressWarnings("GroovyVariableNotAssigned")
     def "inner list objects closure should return the object"() {
         given:
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<Bar> bars
             }
 
-            @DSLConfig
+            @DSL
             class Bar {
                 String name
             }
@@ -442,14 +429,14 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<Bar> bars
             }
 
-            @DSLConfig(key="name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 String url
             }
         ''')
@@ -467,19 +454,20 @@ class TransformSpec extends AbstractDSLSpec {
         instance.bars[1].url == "2"
     }
 
+    @SuppressWarnings("GroovyVariableNotAssigned")
     def "inner list objects closure with named objects should return the created object"() {
         given:
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<Bar> bars
             }
 
-            @DSLConfig(key="name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 String url
             }
         ''')
@@ -502,14 +490,14 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<Bar> bars
             }
 
-            @DSLConfig(key="name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 String url
             }
         ''')
@@ -532,9 +520,9 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
-                @DSLField
+                @Field
                 String name
             }
         ''')
@@ -548,7 +536,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<String> values
                 Map<String, String> fields
@@ -569,7 +557,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<String> values = ['Bla']
                 Map<String, String> fields = [bla: "blub"]
@@ -588,7 +576,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<String> values
             }
@@ -618,9 +606,9 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
-                @DSLField(element="more")
+                @Field(element="more")
                 List<String> values
             }
         ''')
@@ -638,7 +626,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<String> something
             }
@@ -657,7 +645,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List values
             }
@@ -672,7 +660,7 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Map<String, String> values
             }
@@ -702,12 +690,12 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Map<String, Bar> bars
             }
 
-            @DSLConfig
+            @DSL
             class Bar {
                 String name
             }
@@ -722,14 +710,14 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Map<String, Bar> bars
             }
 
-            @DSLConfig(key="name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 String url
             }
         ''')
@@ -745,19 +733,20 @@ class TransformSpec extends AbstractDSLSpec {
         instance.bars.Klaus.url == "2"
     }
 
+    @SuppressWarnings("GroovyVariableNotAssigned")
     def "creation of inner objects in map should return the create object"() {
         given:
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Map<String, Bar> bars
             }
 
-            @DSLConfig(key="name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 String url
             }
         ''')
@@ -780,12 +769,12 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 List<Bar> bars
             }
 
-            @DSLConfig
+            @DSL
             class Bar {
                 String url
             }
@@ -809,14 +798,14 @@ class TransformSpec extends AbstractDSLSpec {
         createInstance('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Map<String, Bar> bars
             }
 
-            @DSLConfig(key = "name")
+            @DSL
             class Bar {
-                String name
+                @Key String name
                 String url
             }
         ''')
@@ -838,7 +827,7 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 String name
             }
@@ -853,14 +842,14 @@ class TransformSpec extends AbstractDSLSpec {
         createClass('''
             package pk
 
-            @DSLConfig
+            @DSL
             class Foo {
                 Bar bar
             }
 
-            @DSLConfig(owner = "owner")
+            @DSL
             class Bar {
-                Foo owner
+                @Owner Foo owner
             }
         ''')
 
@@ -872,7 +861,24 @@ class TransformSpec extends AbstractDSLSpec {
 
         then:
         notThrown(StackOverflowError)
-
     }
+
+    def "error: more than one key"() {
+        when:
+        createClass('''
+            package pk
+
+            @DSL
+            class Foo {
+                @Key String name
+                @Key String name2
+            }
+        ''')
+
+        then:
+        thrown(MultipleCompilationErrorsException)
+    }
+
+
 
 }