+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/model/impl/User3.groovy
@@ -9,11 +9,3 @@ def c = Config.create {
 
 }
 
-Bla {
-
-    options {
-
-    }
-
-
-}